Add remaining stats
This is now able to generate all the applicable stats from nytbee.com, except from our full puzzle set.

There are a ton of obvious outliers to clip based on this analysis.

The chunking logic here is slightly wrong as it doesn't consider the possibility of sparse data.  It's good enough for now.

Also created a utils.dart file.  Bleh.